Restructure directory layout
Move library code into pkg/ and move binaries into their own
subdirectory in cmd/
